TeamsChatBot/requirements.txt
openai
botbuilder-integration-aiohttp>=4.14.0
